fix collections page
