+++ src/clientToServer/fetch/fetchRecipeCollections.ts
@@ -1,11 +1,17 @@
 import type { Recipe } from "../types";
 
 export const fetchRecipeCollections = async (): Promise<Recipe[]> => {
-  const response = await fetch(`/api/user/collections`);
+  try {
+    const response = await fetch(`/api/user/collections`);
 
-  if (!response.ok) {
-    throw new Error("Network response was not ok");
+    if (!response.ok) {
+      throw new Error("Network response was not ok");
+    }
+
+    return response.json();
+
+  } catch (error) {
+    console.error("Failed to fetch recipe collections:", error);
+    throw error;
   }
-
-  return response.json();
 };
