+++ src/components/RecipeCard/index.ts
@@ -1,2 +1,3 @@
 export * from "./RecipeCard";
 export * from "./RecipeCard.types";
+export * from "./RecipeCard.styles"
