+++ src/pages/collections.tsx
@@ -1,28 +1,57 @@
 import * as React from "react";
-import { useSession } from "next-auth/react";
-
-import { useFetchUser } from "src/clientToServer/fetch/useFetchUser";
+import { useQuery } from "@tanstack/react-query";
+import { fetchRecipeCollections } from "src/clientToServer";
+import { useStyles } from "../components/RecipeGallery/RecipeGallery.styles";
+import { RecipeCard } from "../components/RecipeCard";
+import { RecipeProvider } from "../context/RecipeProvider/RecipeProvider";
 
 export default function CollectionsPage() {
-  const { data: session, status } = useSession();
-  const { user } = useFetchUser(session?.user?.id);
+  const {
+    data: recipes,
+    isLoading,
+    error,
+  } = useQuery({
+    queryKey: ["collections"],
+    queryFn: () => fetchRecipeCollections(),
+  });
+
+  const styles = useStyles();
 
   return (
-    <div>
+    <RecipeProvider>
       <h1>Collections</h1>
 
-      {status === "loading" && <p>Loading session...</p>}
-      {status === "unauthenticated" && <p>You are not signed in.</p>}
-
-      {status === "authenticated" && (
-        <>
-          <h2>Your Session Data:</h2>
-          <pre>{JSON.stringify(session, null, 2)}</pre>
-
-          <h2>Your User Data:</h2>
-          <pre>{JSON.stringify(user, null, 2)}</pre>
-        </>
-      )}
-    </div>
+      <div className={styles.grid}>
+        {recipes?.map((recipe, index) => {
+          return (
+            <div
+              key={recipe._id}
+              className={`${styles.fadeIn} ${styles.cardWrapper}`}
+              style={
+                {
+                  "--fadeInDelay": `${Math.min(index * 0.1, 0.3)}s`,
+                } as React.CSSProperties
+              }
+            >
+              <RecipeCard
+                title={recipe?.title}
+                id={recipe?._id}
+                emoji={recipe?.emoji || ""}
+                createdDate={
+                  recipe?.createdAt &&
+                  new Date(recipe?.createdAt).toLocaleDateString(undefined, {
+                    month: "short",
+                    day: "numeric",
+                    year: "numeric",
+                  })
+                }
+                imageSrc={recipe?.imageURL}
+                tags={recipe?.tags}
+              />
+            </div>
+          );
+        })}
+      </div>
+    </RecipeProvider>
   );
 }
